backend/requirements.txt
django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
